backend/pyproject.toml
[project]
name = "buzzerminds-backend"
version = "0.1.0"
description = "BuzzerMinds FastAPI backend"
readme = "README.md"
authors = [
    { name = "Youssef Awad", email = "youssef@ctf.ae" }
]
requires-python = ">=3.14"
dependencies = [
    "alembic>=1.17.1",
    "asyncpg>=0.30.0",
    "elevenlabs>=2.20.0",
    "fastapi>=0.135.1",
    "httpx[http2]>=0.28.1",
    "pydantic-ai-slim[openrouter]>=1.0.5",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.13.1",
    "python-multipart>=0.0.22",
    "pyyaml>=6.0.3",
    "sqlmodel>=0.0.24",
    "uvicorn[standard]>=0.41.0",
    "prometheus-client>=0.22.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.5",
]

[project.scripts]
buzzerminds-backend = "buzzerminds_backend:main"

[tool.pytest.ini_options]
pythonpath = ["src"]

[tool.ruff]
line-length = 100

[build-system]
requires = ["uv_build>=0.10.8,<0.11.0"]
build-backend = "uv_build"
